@Gazler 2020-06-10
I don't think we'll run into any of the issues listed in https://github.com/mzgoddard/hard-source-webpack-plugin#troubleshooting but we could disable this plugin for production mode to be safe.