@joshchernoff 2020-06-22
> I don't think we'll run into any of the issues listed in https://github.com/mzgoddard/hard-source-webpack-plugin#troubleshooting but we could disable this plugin for production mode to be safe.
> 
> ```diff
> -    plugins: [
> -      new HardSourceWebpackPlugin(),
> -      new MiniCssExtractPlugin({ filename: '../css/[name].css' }),
> -      new CopyWebpackPlugin([{ from: 'static/', to: '../' }])
> -    ]
> +    plugins: []
> +      .concat(devMode ? [new HardSourceWebpackPlugin()] : [])
> +      .concat([
> +        new MiniCssExtractPlugin({ filename: '../css/app.css' }),
> +        new CopyWebpackPlugin([{ from: 'static/', to: '../' }])
> +      ])
> ```

This is a great reason why to have separate webpack configs per env.